+++ src/main/java/net/yxiao233/ifeu/common/recipe/DragonStarGeneratorRecipe.java
@@ -1,6 +1,5 @@
 package net.yxiao233.ifeu.common.recipe;
 
-import com.hrznstudio.titanium.component.energy.EnergyStorageComponent;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
@@ -8,19 +7,14 @@ import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementRequirements;
 import net.minecraft.advancements.AdvancementRewards;
 import net.minecraft.advancements.critereon.RecipeUnlockedTrigger;
-import net.minecraft.core.HolderLookup;
 import net.minecraft.data.recipes.RecipeOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
-import net.minecraft.world.level.Level;
-import net.neoforged.neoforge.common.conditions.ICondition;
-import net.neoforged.neoforge.common.conditions.ModLoadedCondition;
-import net.neoforged.neoforge.items.IItemHandler;
-import net.yxiao233.ifeu.common.registry.ModBlocks;
+import net.yxiao233.ifeu.api.recipe.ItemGeneratorSerializableRecipe;
 import net.yxiao233.ifeu.common.registry.ModRecipes;
 
-public class DragonStarGeneratorRecipe implements Recipe<CraftingInput> {
+public class DragonStarGeneratorRecipe extends ItemGeneratorSerializableRecipe {
     public static final MapCodec<DragonStarGeneratorRecipe> CODEC = RecordCodecBuilder.mapCodec((in) -> {
         return in.group(ItemStack.CODEC.fieldOf("input").forGetter((o) -> {
             return o.input;
@@ -33,11 +27,14 @@ public class DragonStarGeneratorRecipe implements Recipe<CraftingInput> {
     public ItemStack input;
     public int time;
     public int powerPerTick;
-    public DragonStarGeneratorRecipe(){}
-    public DragonStarGeneratorRecipe(ItemStack input, int time, int powerPerTick) {
-        this.input = input;
-        this.time = time;
-        this.powerPerTick = powerPerTick;
+
+    public DragonStarGeneratorRecipe(ItemStack input, int progressTime, int powerPerTick) {
+        super(input, progressTime, powerPerTick);
+    }
+
+    @Override
+    public boolean isOnlyForPreview() {
+        return true;
     }
 
     public static void createRecipe(RecipeOutput recipeOutput, String name, DragonStarGeneratorRecipe recipe) {
@@ -48,36 +45,6 @@ public class DragonStarGeneratorRecipe implements Recipe<CraftingInput> {
 
     public static ResourceLocation generateRL(String key) {
         return ResourceLocation.fromNamespaceAndPath("ifeu", "dragon_star_generator/" + key);
-    }
-
-
-    public boolean matches(IItemHandler itemHandler, EnergyStorageComponent<?> energyStorageComponent) {
-        return itemHandler.getStackInSlot(0).is(input.getItem()) && energyStorageComponent.getEnergyStored() <= energyStorageComponent.getMaxEnergyStored();
-    }
-
-    @Override
-    public boolean matches(CraftingInput craftingInput, Level level) {
-        return false;
-    }
-
-    @Override
-    public ItemStack assemble(CraftingInput craftingInput, HolderLookup.Provider provider) {
-        return null;
-    }
-
-    @Override
-    public boolean canCraftInDimensions(int i, int i1) {
-        return false;
-    }
-
-    @Override
-    public ItemStack getResultItem(HolderLookup.Provider provider) {
-        return null;
-    }
-
-    @Override
-    public ItemStack getToastSymbol() {
-        return new ItemStack(ModBlocks.DRAGON_STAR_GENERATOR);
     }
 
     @Override