+++ src/main/java/net/yxiao233/ifeu/common/registry/ModBlocks.java
@@ -20,6 +20,7 @@ public class ModBlocks implements IModule {
     public static BlockWithTile WEATHER_CONTROLLER;
     public static BlockWithTile TIME_CONTROLLER;
     public static BlockWithTile FLUID_CRAFTING_TABLE;
+    public static BlockWithTile DRAGON_GENERATOR;
     @Override
     public void generateFeatures(DeferredRegistryHelper helper) {
         INFUSER = deferredRegistry(helper,"infuser",InfuserBlock::new);
@@ -30,6 +31,7 @@ public class ModBlocks implements IModule {
         WEATHER_CONTROLLER = deferredRegistry(helper,"weather_controller",WeatherControllerBlock::new);
         TIME_CONTROLLER = deferredRegistry(helper,"time_controller",TimeControllerBlock::new);
         FLUID_CRAFTING_TABLE = deferredRegistry(helper,"fluid_crafting_table",FluidCraftingTableBlock::new);
+        DRAGON_GENERATOR = deferredRegistry(helper,"dragon_generator",DragonGeneratorBlock::new);
     }
 
     private BlockWithTile deferredRegistry(DeferredRegistryHelper helper, String name, Supplier<BasicTileBlock<?>> blockSupplier){
@@ -47,6 +49,7 @@ public class ModBlocks implements IModule {
         blocks.add(WEATHER_CONTROLLER.getBlock());
         blocks.add(TIME_CONTROLLER.getBlock());
         blocks.add(FLUID_CRAFTING_TABLE.getBlock());
+        blocks.add(DRAGON_GENERATOR.getBlock());
 
         return blocks;
     }