+++ changelog.md
@@ -18,6 +18,11 @@
 ### v1.1.002
 - 为工业先锋溶解成形机添加了jei一键转移配方的支持
 
+### v1.1.003
+- 添加了龙发电机
+- 修改了龙之星发电机的默认发电量
+- 调整了部分源代码结构
+
 ## en_us
 ### v1.0.0
 - init
@@ -35,3 +40,8 @@
 
 ### v1.1.002
 - Added support for JEI one key transfer recipe to the dissolution chamber
+
+### v1.1.003
+- Added Dragon Generator
+- Modified the default power generation of the Dragon Star generator
+- Adjusted some of the source code structure